Moved UserTestSuite and deleted domain package in tests. Modified: Order and Product Exception, CartMapper and UserMapper
