+++ src/main/java/com/kodilla/ecommercee/controller/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
 package com.kodilla.ecommercee.controller.exception;
 
-public class ProductNotFoundException extends Exception{
+public class ProductNotFoundException extends RuntimeException{
 }
